docker-distribution: fix build for 386
Issue: LIN1025-1486

The build error could be reproduced by using qemux86 as the MACHINE
or using qemux86-64 as the MACHINE and then running 'bitbake
lib32-docker-distribution'.

The root cause is the wrong file (s3_64bit.go) is imported. Fix the
tag matching to ensure s3_32bit.go is imported in case of 386.

(LOCAL REV: NOT UPSTREAM) -- sent to meta-virt on 20250529

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>